Refactoring to run kittydar offline on backend
The thought occurred to me: since the images are being pumped into MongoDB there is nothing stopping
me from post-processing them after they've been written.  So, I modified the MongoDB document structure
a bit to include a flag indicating whether or not the image has been processed yet in addition to the
flag indicating whether or not there was a kitty in the image.  Then, I lowered the max number of threads
used in processing to give more priority to image acquisition and ramped up the framerate.  It still only
cranks out 2-3 images a second, but it's good enough.

I'm writing a utility on the database backend which will pull the images out of MongoDB using a query
on the `kittydar-processed` field, and will periodically query and step through each image, running them
through kittydar.  It will update each document once processing has finished.  Over time, every image can
be processed this way and you can filter out the unprocessed images on the frontend.